+++ packages/gamut-labs/src/GlobalHeader/index.tsx
@@ -3,9 +3,13 @@ import { themed } from '@codecademy/gamut-styles';
 import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import cx from 'classnames';
-import React, { useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 
 import { AppHeader, AppHeaderMobile } from '..';
+import {
+  AppHeaderItem,
+  isAppHeaderItemWithHref,
+} from '../AppHeader/AppHeaderElements/types';
 import {
   FormattedAppHeaderItems,
   FormattedMobileAppHeaderItems,
@@ -102,6 +106,8 @@ const StyledBox = styled(Box)`
 `;
 
 export const GlobalHeader: React.FC<GlobalHeaderProps> = (props) => {
+  const { action, onLinkAction, renderSearch, children } = props;
+
   const [isInHeaderRegion, setIsInHeaderRegion] = useState(true);
 
   // it is not recommended to replicate this logic in other components unless absolutely necessary, as it is
@@ -116,6 +122,14 @@ export const GlobalHeader: React.FC<GlobalHeaderProps> = (props) => {
 
   const theme = useTheme();
 
+  const combinedAction = useCallback(
+    (event: React.MouseEvent, item: AppHeaderItem) => {
+      action(event, item);
+      if (isAppHeaderItemWithHref(item)) onLinkAction?.(event, item);
+    },
+    [action, onLinkAction]
+  );
+
   const headerClasses = cx(
     styles.stickyHeader,
     isInHeaderRegion && styles.transitionFadeOut
@@ -129,7 +143,7 @@ export const GlobalHeader: React.FC<GlobalHeaderProps> = (props) => {
         className={headerClasses}
       >
         <AppHeader
-          action={props.action}
+          action={combinedAction}
           items={getAppHeaderItems(props)}
           redirectParam={
             props.type === 'anon' ? props.redirectParam : undefined
@@ -142,15 +156,15 @@ export const GlobalHeader: React.FC<GlobalHeaderProps> = (props) => {
         className={headerClasses}
       >
         <AppHeaderMobile
-          action={props.action}
+          action={combinedAction}
           items={getMobileAppHeaderItems(props)}
-          renderSearch={props.renderSearch?.mobile}
+          renderSearch={renderSearch?.mobile}
           redirectParam={
             props.type === 'anon' ? props.redirectParam : undefined
           }
         />
       </Box>
-      {props.children}
+      {children}
     </StyledBox>
   );
 };
